user_service/employee_one/requirements/base.txt
# Django
django==1.11.7

# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Forms
django-crispy-forms==1.7.0

# Models
django-model-utils==3.0.0

# Images
Pillow==4.3.0

# Password storage
argon2-cffi==16.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0


# Python-Mysql Database Adapter
mysqlclient==1.3.12

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis>=2.10.5

# DRF
djangorestframework==3.7.3
markdown==2.6.9
django-filter==1.1.0





# Your custom requirements go here
